Fixed TF logging in train_naive

+++ transformer/decoder_clf.py
@@ -12,7 +12,7 @@ class DecoderClf(nn.Module):
     def forward(self, x: Tensor, attention_mask: Tensor) -> Tensor:
         x = self.lm(x, None, attention_mask)
         # get the last token per sample in batch
-        last_token_idx = attention_mask.sum(dim=1) - 1
+        last_token_idx = (attention_mask.sum(dim=1) - 1).long()
         x = x.gather(1, last_token_idx.unsqueeze(1).unsqueeze(2).expand(-1, -1, x.size(2))).squeeze(1)
         return self.fc(x)
